Move Cache and related classes to internal package
