+++ instrumentation-api/src/main/java/io/opentelemetry/instrumentation/api/internal/cache/concurrentlinkedhashmap/Weigher.java
@@ -5,7 +5,7 @@
 
 // Includes work from:
 /*
- * Copyright 2012 Google Inc. All Rights Reserved.
+ * Copyright 2010 Google Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,28 +20,27 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.instrumentation.api.cache.internal.concurrentlinkedhashmap;
+package io.opentelemetry.instrumentation.api.internal.cache.concurrentlinkedhashmap;
 
 import javax.annotation.concurrent.ThreadSafe;
 
 /**
- * A class that can determine the weight of an entry. The total weight threshold is used to
- * determine when an eviction is required.
+ * A class that can determine the weight of a value. The total weight threshold is used to determine
+ * when an eviction is required.
  *
  * @author ben.manes@gmail.com (Ben Manes)
  * @see <a href="http://code.google.com/p/concurrentlinkedhashmap/">
  *     http://code.google.com/p/concurrentlinkedhashmap/</a>
  */
 @ThreadSafe
-public interface EntryWeigher<K, V> {
+public interface Weigher<V> {
 
   /**
-   * Measures an entry's weight to determine how many units of capacity that the key and value
-   * consumes. An entry must consume a minimum of one unit.
+   * Measures an object's weight to determine how many units of capacity that the value consumes. A
+   * value must consume a minimum of one unit.
    *
-   * @param key the key to weigh
-   * @param value the value to weigh
-   * @return the entry's weight
+   * @param value the object to weigh
+   * @return the object's weight
    */
-  int weightOf(K key, V value);
+  int weightOf(V value);
 }
